pyproject.toml
[project]
name = "video-subtitle-merger"
authors = [
    {name="Chinh Dinh", email="dinhminhchinh3357@gmail.com"}
]
version = "0.1.0"
description = "Merge subtitles into a video"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "moviepy>=2.1.2",
    "streamlit>=1.42.0",
    "webvtt-py>=0.5.1",
]
